+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Added tooltip on badge
+* Removed underline on hover of badge
+
 ### 1.0.0: June 05th, 2020
 * Plugin first version with realtime badge occupation
 * Cron getting info from webservice every 15 minutes